Move various components into their own files
