+++ src/Components/Instructions.jsx
@@ -1,0 +1,39 @@
+import React from 'react';
+import Square from './Square';
+
+function Instructions() {
+  return (
+    <div>
+      <div className="subtitle">
+        Welcome to Wurmdle! Try to guess the Pokemon based on its base stats!
+        You have five guesses. Adjust the slider to change which generations to
+        play with. Report issues
+        {' '}
+        <a href="https://github.com/veliakiner/wurmdle/issues">here</a>
+        .
+      </div>
+      <div className="key">
+        <div className="key-elem">Key:</div>
+        <div className="keys">
+          <span className="key-elem">
+            <Square key="toolow" value="0-" />
+            {' '}
+            Too low
+          </span>
+          <span className="key-elem">
+            <Square key="toohigh" value="999+" />
+            {' '}
+            Too high
+          </span>
+          <span className="key-elem">
+            <Square key="correct" value="100=" />
+            {' '}
+            Correct
+          </span>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default Instructions;
